Update README.md
Returning to this project to add the details now that the assessment marks have been fully confirmed. The project has been dormant following my realisation that sharing my assignment with the world prior to this was inappropriate - I was far too eager to engage in the open-science possibilities of RStudio initially.